+++ src/main/java/com/nuguna/freview/customer/dto/request/CustomerMyApplyInfosRetrieveRequestDTO.java
@@ -13,6 +13,6 @@ public class CustomerMyApplyInfosRetrieveRequestDTO {
 
   @Min(1)
   @NotNull
-  private Long targetPage;
+  private Integer targetPage;
 
 }
